requirements.txt
# UI & Environment
streamlit
python-dotenv

# Core LangChain (Matched to latest available)
langchain
langchain-core
langchain-community
langchain-huggingface

# AI & Embeddings
sentence-transformers
huggingface-hub
transformers
torch

# Vector Database & Utilities
faiss-cpu
pypdf
numpy
pandas